+++ src/solve_simple.rs
@@ -6,7 +6,7 @@ use itertools::Itertools;
 pub fn solve_basic(cnf: &CnfFormula) -> Option<Assignment> {
     // Literally iterate through every possible assignment.
     std::iter::repeat([Some(Val::False), Some(Val::True)])
-        .take(cnf.num_vars as usize)
+        .take(cnf.num_vars)
         .multi_cartesian_product()
         .map(|v| Assignment::from_vector(v.to_vec()))
         .find(|assignment| check_assignment(cnf, assignment))
